Logs/metagpt/part3-2/instruction48/student_evaluation_automation/requirements.txt
Flask==3.0.3
gspread==5.12.0
google-api-python-client==2.126.0
google-auth==2.29.0
google-auth-oauthlib==1.2.0
openai==1.33.0
pdfkit==1.0.0
simple-salesforce==1.12.4
requests==2.32.3
python-dotenv==1.0.1
gunicorn==22.0.0
Werkzeug==3.0.3
email-validator==2.1.1